fix: Fix 200 OK when deleting a non-exist job

+++ src/main/java/com/example/jobapp/job/impl/JobServiceImpl.java
@@ -35,12 +35,11 @@ public class JobServiceImpl implements JobService {
 
     @Override
     public boolean deleteJobById(Long id) {
-        try {
+        if (jobRepository.existsById(id)) {
             jobRepository.deleteById(id);
             return true;
-        } catch (Exception e) {
-            return false;
         }
+        return false;
     }
 
     @Override
